pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "paramview"
version = "0.3.0"
description = "GUI for viewing the contents of a ParamDB database."
authors = ["Alex Hadley <contact@alexhadley.net>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/PainterQubits/paramview"
include = ["paramview/static/**/*"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
flask-socketio = "^5.3.6"
eventlet = "^0.33.3"
watchdog = "^3.0.0"
paramdb = "^0.10.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
flake8 = "^6.1.0"
pylint = "^3.0.2"
black = "^23.11.0"
pytest = "^7.4.3"
pytest-playwright = "^0.4.3"
freezegun = "1.2.2"
requests = "^2.31.0"
sqlalchemy = "^2.0.23"
astropy = "^6.0.0"

[tool.poetry.scripts]
paramview = "paramview._cli:main"

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
base_url = "http://localhost:5050"
filterwarnings = ["ignore::DeprecationWarning:eventlet.support.greenlets"]
